personalização da interface de cadastro de professor e correção ao excluir itens que contenha chave estrangeira cadastrada em outra tabela

+++ uAluno.pas
@@ -69,6 +69,14 @@ begin
   query := TFDQuery.Create(nil);
   try
     query.Connection := connection;
+
+    query.SQL.Text :='SELECT COUNT(*) AS total FROM matricula where codigo_aluno = :id_aluno';
+    query.ParamByName('id_aluno').AsInteger := self.getCodigo;
+
+    if query.FieldByName('total').AsInteger > 0 then begin
+      raise Exception.Create('o aluno selecionado esta cadastrado ha uma matricula,tente editar a matricula ou exclui-la antes de remover esse aluno');
+    end;
+
     query.SQL.Text := 'DELETE FROM aluno WHERE CODIGO = ' + IntToStr(self.getCodigo);
     query.ExecSQL;
   finally
