+++ uDisciplina.pas
@@ -63,6 +63,14 @@ begin
   query := TFDQuery.Create(nil);
   try
     query.Connection := connection;
+
+    query.SQL.Text :='SELECT COUNT(*) AS total FROM turma where codigo_disciplina = :id_disciplina';
+    query.ParamByName('id_disciplina').AsInteger := self.getCodigo;
+
+    if query.FieldByName('total').AsInteger > 0 then begin
+      raise Exception.Create('A disciplina selecionada esta cadastrada ha uma turma,tente editar a turma ou exclui-la antes de remover essa matricula');
+    end;
+
     query.SQL.Text := 'DELETE FROM disciplina WHERE CODIGO = ' + IntToStr(self.getCodigo);
     query.ExecSQL;
   finally
